Creating AND, OR clause. NOT will be handle at IN clause
