+++ sqlNice/columnnice.py
@@ -3,6 +3,7 @@
 class ColumnNice(object):
     def __init__(self, column_name):
         self.name = column_name
+        self.operation = None
 
     def __str__(self):
         return self.name
@@ -13,7 +14,11 @@ class ColumnNice(object):
         :param other:
         :return: string of sql query
         """
-        return str(self) + ' < ' + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' < ' + str(other)
+        return self
 
     def __le__(self, other):
         """
@@ -21,7 +26,11 @@ class ColumnNice(object):
         :param other:
         :return:
         """
-        return str(self) + ' <= ' + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' <= ' + str(other)
+        return self
 
     def __eq__(self, other):
         """
@@ -29,7 +38,11 @@ class ColumnNice(object):
         :param other:
         :return:
         """
-        return str(self) + ' == ' + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' == ' + str(other)
+        return self
 
     def __ne__(self, other):
         """
@@ -37,7 +50,11 @@ class ColumnNice(object):
         :param other:
         :return:
         """
-        return str(self) + " != " + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' != ' + str(other)
+        return self
 
     def __gt__(self, other):
         """
@@ -45,7 +62,11 @@ class ColumnNice(object):
         :param other:
         :return:
         """
-        return str(self) + ' > ' + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' > ' + str(other)
+        return self
 
     def __ge__(self, other):
         """
@@ -53,7 +74,11 @@ class ColumnNice(object):
         :param other:
         :return:
         """
-        return str(self) + ' >= ' + str(other)
+        if type(other) == str:
+            other = "\"" + other + "\""
+
+        self.operation = str(self) + ' >= ' + str(other)
+        return self
 
     def __contains__(self, item):
         """
@@ -62,6 +87,14 @@ class ColumnNice(object):
         :return:
         """
         pass
+
+    def __and__(self, other):
+        self.operation += ' AND ' + other.operation
+        return self
+
+    def __or__(self, other):
+        self.operation += ' OR ' + other.operation
+        return self
 
     def like(self, other):
         pass
